Extend deployer to handle linked factories
