+++ pytest_ethereum/deployer.py
@@ -1,6 +1,7 @@
 from eth_utils import to_canonical_address
-
 from ethpm import Package
+
+from pytest_ethereum.exceptions import DeployerError
 
 
 class Deployer:
@@ -13,8 +14,11 @@ class Deployer:
 
     def deploy(self, contract_type, *args):
         factory = self.package.get_contract_factory(contract_type)
-        if factory.has_linkable_bytecode():
-            raise TypeError("don't support link refs yet.")
+        if factory.has_linkable_bytecode() and not factory.is_bytecode_linked:
+            raise DeployerError(
+                "Unable to deploy an unlinked factory. "
+                "Please create a Deployer with a linked factory instance. "
+            )
         tx_hash = factory.constructor(*args).transact()
         tx_receipt = self.package.w3.eth.waitForTransactionReceipt(tx_hash)
         return self.package, to_canonical_address(tx_receipt.contractAddress)
