+++ pytest_ethereum/exceptions.py
@@ -1,0 +1,14 @@
+class PytestEthereumError(Exception):
+    """
+    Base class for all Pytest-Ethereum errors.
+    """
+
+    pass
+
+
+class DeployerError(PytestEthereumError):
+    """
+    Raised when the Deployer is unable to deploy a contract type.
+    """
+
+    pass
