+++ setup.py
@@ -48,8 +48,7 @@ setup(
     include_package_data=True,
     install_requires=[
         "eth-utils>=1,<2",
-        "eth-tester==0.1.0b26",
-        "ethpm>=0.1.1,<1",
+        "ethpm>=0.1.3,<2",
         "web3[tester]==4.4.1",
         "vyper>=0.1.0b2,<1",
     ],
