+++ tests/core/test_import.py
@@ -1,0 +1,2 @@
+def test_import():
+    import pytest_ethereum  # noqa: F401
