+++ tests/manifests/safe_math_manifest.json
