+++ tests/manifests/standard_token_manifest.json
