+++ tox.ini
@@ -38,5 +38,5 @@ basepython=python
 extras=lint
 commands=
     flake8 {toxinidir}/pytest_ethereum {toxinidir}/tests
-	black --check --diff {toxinidir}/pytest_ethereum/ --check --diff {toxinidir}/tests/
+    black --check --diff {toxinidir}/pytest_ethereum/ --check --diff {toxinidir}/tests/
     isort --recursive {toxinidir}/pytest_ethereum {toxinidir}/tests
